Move axe work into parent class
