+++ lighthouse-core/audits/accessibility/aria-allowed-attr.js
@@ -1,0 +1,40 @@
+/**
+ * @license
+ * Copyright 2016 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+'use strict';
+
+/**
+ * @fileoverview Ensures ARIA attributes are allowed for an element's role.
+ */
+
+const AxeAudit = require('./axe-audit');
+
+class ARIAAllowedAttr extends AxeAudit {
+  /**
+   * @return {!AuditMeta}
+   */
+  static get meta() {
+    return {
+      category: 'Accessibility',
+      name: 'aria-allowed-attr',
+      description: 'Element aria-* roles are valid',
+      requiredArtifacts: ['Accessibility']
+    };
+  }
+}
+
+module.exports = ARIAAllowedAttr;
