+++ lighthouse-core/audits/accessibility/aria-valid-attr-value.js
@@ -21,11 +21,9 @@
  * @fileoverview Ensures all ARIA attributes have valid values.
  */
 
-const Audit = require('./audit');
-const A11yHelpers = require('../lib/a11y-helpers');
-const Formatter = require('../formatters/formatter');
+const AxeAudit = require('./axe-audit');
 
-class ARIAValidAttr extends Audit {
+class ARIAValidAttr extends AxeAudit {
   /**
    * @return {!AuditMeta}
    */
@@ -36,24 +34,6 @@ class ARIAValidAttr extends Audit {
       description: 'Element aria-* attributes have valid values',
       requiredArtifacts: ['Accessibility']
     };
-  }
-
-  /**
-   * @param {!Artifacts} artifacts
-   * @return {!AuditResult}
-   */
-  static audit(artifacts) {
-    const violations = artifacts.Accessibility.violations || [];
-    const rule = violations.find(result => result.id === 'aria-valid-attr-value');
-
-    return ARIAValidAttr.generateAuditResult({
-      rawValue: typeof rule === 'undefined',
-      debugString: A11yHelpers.createDebugString(rule),
-      extendedInfo: {
-        formatter: Formatter.SUPPORTED_FORMATS.ACCESSIBILITY,
-        value: rule
-      }
-    });
   }
 }
