+++ lighthouse-core/config/default.json
@@ -68,15 +68,15 @@
     "manifest-short-name-length",
     "manifest-start-url",
     "theme-color-meta",
-    "aria-valid-attr",
-    "aria-valid-attr-value",
-    "aria-allowed-attr",
-    "aria-required-attr",
-    "color-contrast",
-    "image-alt",
-    "label",
-    "tabindex",
     "content-width",
+    "accessibility/aria-valid-attr",
+    "accessibility/aria-valid-attr-value",
+    "accessibility/aria-allowed-attr",
+    "accessibility/aria-required-attr",
+    "accessibility/color-contrast",
+    "accessibility/image-alt",
+    "accessibility/label",
+    "accessibility/tabindex",
     "dobetterweb/external-anchors-use-rel-noopener",
     "dobetterweb/appcache-manifest",
     "dobetterweb/geolocation-on-start",
